refactor: improve overall structure
