+++ package.json
@@ -49,24 +49,25 @@
     "@types/react": "^17.0.33",
     "@types/react-dom": "^17.0.10",
     "@types/webextension-polyfill": "^0.8.1",
-    "@typescript-eslint/eslint-plugin": "^5.1.0",
-    "@typescript-eslint/parser": "^5.1.0",
+    "@typescript-eslint/eslint-plugin": "^5.10.1",
+    "@typescript-eslint/parser": "^5.10.1",
     "archiver": "^5.3.0",
     "autoprefixer": "^10.4.0",
     "concurrently": "^7.0.0",
     "copy-webpack-plugin": "^9.0.1",
     "cross-env": "^7.0.3",
     "css-loader": "^6.5.0",
-    "eslint": "^7.2.0",
-    "eslint-config-airbnb": "18.2.1",
+    "eslint": "^8.8.0",
+    "eslint-config-airbnb": "19.0.4",
+    "eslint-config-next": "^12.0.9",
     "eslint-config-prettier": "^8.3.0",
     "eslint-import-resolver-typescript": "^2.5.0",
-    "eslint-plugin-import": "^2.22.1",
-    "eslint-plugin-import-helpers": "^1.1.0",
-    "eslint-plugin-jsx-a11y": "^6.4.1",
+    "eslint-plugin-import": "^2.25.3",
+    "eslint-plugin-import-helpers": "^1.2.1",
+    "eslint-plugin-jsx-a11y": "^6.5.1",
     "eslint-plugin-prettier": "^4.0.0",
-    "eslint-plugin-react": "^7.21.5",
-    "eslint-plugin-react-hooks": "^1.7.0",
+    "eslint-plugin-react": "^7.28.0",
+    "eslint-plugin-react-hooks": "^4.3.0",
     "fs-extra": "^10.0.0",
     "html-webpack-plugin": "^5.5.0",
     "husky": "^7.0.4",
