--- scripts/.eslintrc.json
@@ -1,6 +1,0 @@
-{
-  "rules": {
-    "import/no-extraneous-dependencies": "off",
-    "@typescript-eslint/no-var-requires": "off"
-  }
-}
