+++ src/entries/popup/popup.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import ReactDOM from 'react-dom';
 import 'tailwindcss/tailwind.css'; // eslint-disable-line import/no-extraneous-dependencies
 
-import PopUpPage from '~/pages/PopUpPage';
+import PopUpPage from '~/components/pages/PopUpPage/PopUpPage';
 
 ReactDOM.render(
   <React.StrictMode>
