+++ src/utils/math/math.ts
